Standardize Field Names
Created a basic function and script to standardize field names in terms of:
A) Spelling
B) Order

The function takes a filename and LIST of incorrect fieldnames contained in file, in the proper order.

ex.
StandardizeFieldNames(test.csv, ['twetid', 'usrid', 'twet',...]

Can be run in three ways:
1) Edit the script vars and run yourself without an argument
2) Call the function from elsewhere
3) run on command line with argument 1 = filename and 2= list,of,wrong,fieldnames,comma,but,no,spaces,spaced.

Lastly, you can test that it works by running the file with the arugment 'test'. This will run the test files I created.

The 'og' file is the 'wrong' file and saved that way. 'test' is the of file, but reordered. 'test_modified' is the final result, and should match 'corrected'.

test Files can be found in data > fieldnameFixFiles.

NOTES:
1) File to modify MUST be in data > fieldnameFixFiles.

This must be done BEFORE data is labeled (I generally do this as I combine the two datasets).