Data Preprocessing
Pipeline prototype up, trying to make it as encapsulated as possible and as automated as possible now. It will require tinkering on a per-file basis, but keeping all things needing change in the top few var assignments.

NOTE:
Larger datafiles will not be pushed, have to add and name them on your own